+++ CHANGELOG.md
@@ -6,7 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Release]
 
-### [1.0] - 2020-01-3
+### [2.0] - 2021-01-04
 #### Added
-- Added library.
-- Added new method to list package names with application names.
+- Added new method to generate application signature.
+- Added method to search for application ID by application name.
+- Added method to get application name based on application ID.
+- Added method to check if an application requests a specific permission.
+- Added method to check if an application is granted a specific permission.
+- Added method to search all applications requesting a specific permission.
+- Added method to search all applications granted a specific permission.
+
+### [1.0] - 2020-12-21
+#### Added
+- Released library publicly.
+- Added new method to list all applications with their IDs.